requirements.txt
rich
python-dotenv
beautifulsoup4
tweepy
openai
anthropic
supabase